requirements-base.txt
# Core FastAPI & Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3

# Database & ORM
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
alembic==1.12.1

# Task Queue
celery==5.3.4
redis==5.0.0
flower==2.0.1

# API & HTTP
requests==2.31.0
httpx==0.25.1

# Logging & Monitoring
python-json-logger==2.0.7
prometheus-client==0.19.0

# Security
PyJWT==2.8.0
passlib==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# Rate Limiting
slowapi==0.1.9

# Utilities
python-dotenv==1.0.0
pytz==2023.3
python-dateutil==2.8.2

# Validation
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
